Refactor: Post

+++ src/posts/post.repository.js
@@ -9,7 +9,7 @@ dotenv.config();
 class PostRepository {
   constructor() {
   }
-  getPostById = async (postId) => {
+   getPostById = async (postId) => {
     try {
       const post = await Posts.findOne({ where: { postId: postId } });
       const images = await Images.findAll({ where: { postId: postId } });
@@ -42,74 +42,56 @@ class PostRepository {
       }
 
       // Extract the information
-      let { content, musicType, musicUrl, musicTitle, latitude, longitude, placeName, tag } = body;
+      let { content, musicType, musicUrl, musicTitle, latitude, longitude, placeName, tag,imageURL } = body;
       const images = files['image'];
       const audio = files['audio'] ? files['audio'][0] : null;
 
-      console.log(audio)
-
-      if (musicType==="2") {
-        switch (musicUrl){
-          case "1":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/1.Alan Walker - Dreamer (BEAUZ & Heleen Remix) [NCS Release].mp3`;
-            musicTitle='Alan Walker - Dreamer (BEAUZ & Heleen Remix) [NCS Release]';
-            break;
-          case "2":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/2.Arcando & Maazel - To Be Loved (feat. Salvo) [NCS Release].mp3`;
-            musicTitle='Arcando & Maazel - To Be Loved (feat. Salvo) [NCS Release]';
-            break;
-          case "3":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/3.AX.EL - In Love With a Ghost [NCS Release].mp3`;
-            musicTitle='AX.EL - In Love With a Ghost [NCS Release]';
-            break;
-          case "4":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/4.Idle Days - Over It [NCS Release].mp3`
-            musicTitle='Idle Days - Over It [NCS Release]'
-            break;
-          case "5":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/5.ROY KNOX - Closer [NCS Release].mp3`
-            musicTitle='ROY KNOX - Closer [NCS Release]'
-            break;
-        }
-      }
+      console.log(imageURL)
 
       if (musicType==="3") {
-        musicTitle='녹음된 음원';
-        musicUrl = audio.location;
+        musicTitle = '녹음된 음원';
+        musicUrl = audio ? audio.location : musicUrl;
       }
 
       // Update the post
 
-      await post.update({ content,
-        musicTitle,
-        musicUrl,
-        latitude,
-        longitude,
-        placeName,
-        musicType,
-        private: false,
-        tag});
 
-      // Delete old images and audio
-      const audioDelete = await Records.findOne({ where: { postId: post.postId }})
-      if(audioDelete) {
-        const audioPath = new URL(audioDelete.dataValues.url).pathname.substr(1)
-        await deleteS3(audioPath)
-        await Records.destroy({where: {postId: postId}});
+      if(audio) {
+        const audioDelete = await Records.findOne({where: {postId: post.postId}})
+        if (audioDelete) {
+          const audioPath = new URL(audioDelete.dataValues.url).pathname.substr(1)
+          await deleteS3(audioPath)
+          await Records.destroy({where: {postId: postId}});
+        }
       }
 
-      const imageDelete = await Images.findAll({ where: { postId: post.postId }})
-      if(imageDelete) {
-        const s3DeletePromises = imageDelete.map((image) => {
+      const imageDelete = await Images.findAll({ where: { postId: post.postId }});
+
+      const urlsToDelete = imageDelete.map(image => image.url); // imageDelete 배열에서 URL만 추출합니다.
+
+      const urlsNotIncluded = imageDelete.filter(image => !imageURL.includes(image.url));
+
+      console.log('urlsToDelete:',urlsToDelete)
+      console.log("urlIsNotIncluded:",urlsNotIncluded)
+      //console.log('imageDelete:',imageDelete)
+      console.log('imageURL:',imageURL)
+
+      if(urlsNotIncluded) {
+        const s3DeletePromises = urlsNotIncluded.map(async (image) => {
           const imagePath = new URL(image.dataValues.url).pathname.substr(1)
+          await Images.destroy({where: {imageId: image.imageId}});
           return deleteS3(imagePath)
         })
-
         await Promise.all(s3DeletePromises)
-
-        await Images.destroy({where: {postId: postId}});
+        //await Images.destroy({where: {imageId: image.imageId}});
       }
 
+      if(images){
       const imagePromises = images.map((image) => {
         return Images.create({ url: image.location, postId: postId, userId: post.dataValues.userId });
       });
 
-      await Promise.all(imagePromises);
+      await Promise.all(imagePromises);}
 
       console.log(audio)
 
@@ -121,12 +103,21 @@ class PostRepository {
         });
       }
 
+      await post.update({ content,
+        musicTitle,
+        musicUrl,
+        latitude,
+        longitude,
+        placeName,
+        musicType,
+        private: false,
+        tag});
+
     } catch (err) {
       console.error(err);
       throw { status: 500, message: 'Error updating post' };
     }
   };
-
 
   deletePost = async (postId) => {
     try {
@@ -174,26 +165,6 @@ class PostRepository {
     let { content, musicType, musicUrl, musicTitle, latitude, longitude, placeName, tag } = body;
     const images = files['image'];
     const audio = files['audio'] ? files['audio'][0] : null;
-
-    if (musicType==="2") {
-      switch (musicUrl){
-        case "1":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/1.Alan Walker - Dreamer (BEAUZ & Heleen Remix) [NCS Release].mp3`
-          musicTitle='Alan Walker - Dreamer (BEAUZ & Heleen Remix) [NCS Release]'
-          break;
-        case "2":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/2.Arcando & Maazel - To Be Loved (feat. Salvo) [NCS Release].mp3`
-          musicTitle='Arcando & Maazel - To Be Loved (feat. Salvo) [NCS Release]'
-          break;
-        case "3":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/3.AX.EL - In Love With a Ghost [NCS Release].mp3`
-          musicTitle='AX.EL - In Love With a Ghost [NCS Release]'
-          break;
-        case "4":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/4.Idle Days - Over It [NCS Release].mp3`
-          musicTitle='Idle Days - Over It [NCS Release]'
-          break;
-        case "5":musicUrl=`${process.env.ORIGIN_BACK}/publicMusic/5.ROY KNOX - Closer [NCS Release].mp3`
-          musicTitle='ROY KNOX - Closer [NCS Release]'
-          break;
-      }
-    }
 
     if (musicType==="3") {
       musicTitle = '녹음된 음원';
